chore: Small restructuring cleanup
* Move all tests to tests/ directory
* Re-export common functionality from ./src/index.ts
* Add helper makefile
* Move instaboot tests to dedicated file
