+++ fixtures/path-logger.php
@@ -1,4 +1,5 @@
 <?php
+
  // Get the requested path
 $path = $_SERVER['REQUEST_URI'];
