--- run.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-DENO_JOBS=8 deno test --allow-all --quiet --parallel test.ts
