+++ src/edge.ts
@@ -1,0 +1,4 @@
+export const EDGE_HEADER_PURGE_INSTANCES = "x-edge-purge-instances";
+export const EDGE_HEADER_JOURNAL_STATUS = "x-edge-instance-journal-status";
+export const HEADER_PURGE_INSTANCES: string = "x-edge-purge-instances";
+export const HEADER_INSTANCE_ID: string = "x-edge-instance-id";
